refactor: extract helpers from common.ts to yamlUtils.ts
