+++ packages/shared/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * as PackageUtils from './packageUtils';
 export * as MiscUtils from './miscUtils';
 export * as RepositoryUtils from './repositoryUtils';
+export * as YamlUtils from './yamlUtils';
